requirements.txt
tensorflow>=2.12
numpy
pandas
matplotlib
scikit-learn
flask
opencv-python
Pillow
